Replace floating point FPS with integer
Former-commit-id: 7ac2a2bb650f861df6f9e7042363595c0c919091

+++ include/IME/core/loop/Engine.h
@@ -88,7 +88,7 @@ namespace IME {
          * @brief Get frames per second (FPS) limit
          * @return FPS limit
          */
-        float getFPSLimit() const;
+        unsigned int getFPSLimit() const;
 
         /**
          * @brief Get the name of the game
